Remove unused members from `PredictionFramework` class definition and add compartments_to_vector() to `ModelCommon.h`.
