+++ src/InferenceFramework.cpp
@@ -3,6 +3,7 @@
 #include "InferenceParameters.h"
 #include "IO.h"
 #include "FittingProcess.h"
+#include "ModelCommon.h"
 #include <functional>
 #include <algorithm>
 
@@ -310,7 +311,7 @@ void InferenceFramework::ModelSelect(EERAModel::particle& outvec, const std::vec
 	outvec.simu_outs = status.simulation;
 	outvec.hospital_death_outs = status.hospital_deaths;
 	outvec.death_outs = status.deaths;
-	outvec.end_comps = compartments_to_vector(status.ends);
+	outvec.end_comps = Model::compartments_to_vector(status.ends);
 }
 
 
@@ -343,21 +344,6 @@ std::discrete_distribution<int> ComputeWeightDistribution(
 	}
 	
 	return std::discrete_distribution<int>(weight_val.begin(), weight_val.end());
-}
-
-std::vector<std::vector<int>> compartments_to_vector(const std::vector<Compartments>& cmps_vec)
-{
-	std::vector<std::vector<int>> _temp;
-
-	for(auto cmps : cmps_vec)
-	{
-		_temp.push_back({cmps.S, cmps.E, cmps.E_t, cmps.I_p,
-						cmps.I_t, cmps.I1, cmps.I2, cmps.I3,
-						cmps.I4, cmps.I_s1, cmps.I_s2, cmps.I_s3,
-						cmps.I_s4, cmps.H, cmps.R, cmps.D});
-	}
-
-	return _temp;
 }
 
 } // namespace Inference
