+++ src/InferenceFramework.h
@@ -122,17 +122,5 @@ std::discrete_distribution<int> ComputeWeightDistribution(
 	const std::vector<EERAModel::particle>& particleList);
 
 
-/**
- * @brief Convert Vector of Compartments struct to a vector of integers
- * 
- * NOTE: This is a temporary function to allow compatibility
- * converts the Compartments struct to a vector of integers
- * 
- * @param cmps_vec Vector of compartments struct containing population per category
- * 
- * @return Vector of population counters
- */
-std::vector<std::vector<int>> compartments_to_vector(const std::vector<Compartments>& cmps_vec);
-
 } // namespace Inference
 } // namespace EERAModel
