+++ src/ModelCommon.h
@@ -101,5 +101,17 @@ std::vector<int> ComputeAgeNums(int shb_id, int Npop, int N_hcw, const InputObse
  */
 std::vector<params> BuildFixedParameters(unsigned int size, params parameters);
 
+/**
+ * @brief Convert Vector of Compartments struct to a vector of integers
+ * 
+ * NOTE: This is a temporary function to allow compatibility
+ * converts the Compartments struct to a vector of integers
+ * 
+ * @param cmps_vec Vector of compartments struct containing population per category
+ * 
+ * @return Vector of population counters
+ */
+std::vector<std::vector<int>> compartments_to_vector(const std::vector<Compartments>& cmps_vec);
+
 } // namespace Model
 } // namespace EERAModel
