+++ src/PredictionFramework.cpp
@@ -77,24 +77,9 @@ void PredictionFramework::Run(std::vector<double> parameterSet, int nSimulationS
     (*log_) << "\n <computation time> " << time_taken << " seconds.\n";
 
     std::vector< std::vector<int> > end_comps;
-	end_comps = compartments_to_vector(status.ends);
+	end_comps = Model::compartments_to_vector(status.ends);
 
     IO::WritePredictionsToFiles(status, end_comps, outDir_, log_);
-}
-
-std::vector<std::vector<int>> compartments_to_vector(const std::vector<Compartments>& cmps_vec)
-{
-	std::vector<std::vector<int>> _temp;
-
-	for(auto cmps : cmps_vec)
-	{
-		_temp.push_back({cmps.S, cmps.E, cmps.E_t, cmps.I_p,
-						cmps.I_t, cmps.I1, cmps.I2, cmps.I3,
-						cmps.I4, cmps.I_s1, cmps.I_s2, cmps.I_s3,
-						cmps.I_s4, cmps.H, cmps.R, cmps.D});
-	}
-
-	return _temp;
 }
 
 } // namespace Prediction
